Move activatePromotion and dismissPromotion from resolvers to actions
These functions are mutation actions, not data resolvers, so they need to be exported from actions.ts to be available via useDispatch.

Changes:
- Moved activatePromotion and dismissPromotion from resolvers.ts to actions.ts
- Updated to use invalidateResolution instead of calling getPromotions directly
- These are now properly available as dispatchable actions via usePromotionActions hook

This fixes the "activatePromotion is not a function" error when trying to use the promotion actions.
