+++ client/data/promotions/resolvers.ts
@@ -36,78 +36,3 @@ export function* getPromotions(): unknown {
 		yield updateErrorForPromotions( e as ApiError );
 	}
 }
-
-/**
- * Activate a promotion.
- *
- * @param {string} identifier The promotion identifier.
- * @param {boolean} acceptTerms Whether to accept the promotion terms.
- */
-export function* activatePromotion(
-	identifier: string,
-	acceptTerms = true
-): unknown {
-	const path = `${ NAMESPACE }/promotions/${ identifier }/activate`;
-
-	try {
-		yield apiFetch( {
-			path,
-			method: 'POST',
-			data: { accept_terms: acceptTerms },
-		} );
-
-		yield controls.dispatch(
-			'core/notices',
-			'createSuccessNotice',
-			__( 'Promotion activated successfully!', 'woocommerce-payments' )
-		);
-
-		// Refresh promotions list after activation.
-		yield getPromotions();
-	} catch ( e ) {
-		yield controls.dispatch(
-			'core/notices',
-			'createErrorNotice',
-			__(
-				'Error activating promotion. Please try again.',
-				'woocommerce-payments'
-			)
-		);
-		yield updateErrorForPromotions( e as ApiError );
-	}
-}
-
-/**
- * Dismiss a promotion.
- *
- * @param {string} identifier The promotion identifier.
- */
-export function* dismissPromotion( identifier: string ): unknown {
-	const path = `${ NAMESPACE }/promotions/${ identifier }/dismiss`;
-
-	try {
-		yield apiFetch( {
-			path,
-			method: 'POST',
-		} );
-
-		yield controls.dispatch(
-			'core/notices',
-			'createSuccessNotice',
-			__( 'Promotion dismissed.', 'woocommerce-payments' )
-		);
-
-		// Refresh promotions list after dismissal.
-		yield getPromotions();
-	} catch ( e ) {
-		yield controls.dispatch(
-			'core/notices',
-			'createErrorNotice',
-			__(
-				'Error dismissing promotion. Please try again.',
-				'woocommerce-payments'
-			)
-		);
-		yield updateErrorForPromotions( e as ApiError );
-	}
-}
